+++ CHANGELOG.md
@@ -1,62 +1,3 @@
-## [1.13.1] - 2025-03-21
-
-### Fixed
-
-- Fixed an issue where threads would not load properly (Daniel D’Aquino)
-
-
-[1.13.1]: https://github.com/damus-io/damus/releases/tag/v1.13.1
-
-
-## [1.13] - 2025-03-14
-
-### Added
-
-- Added local persistence of note drafts (Daniel D’Aquino)
-- Added user-friendly error view for errors around the app that would not fit in other places (Daniel D’Aquino)
-- Coinos connection button in Wallet view (ericholguin)
-- Added Alby Go to mobile wallets selection menu (Tomek ⚡ K)
-- Minor accessibility improvements around picture editing and onboarding (Daniel D’Aquino)
-- Profile image cropping tools (Daniel D’Aquino)
-- Added Conversations tab to profiles (Terry Yiu)
-- Added profile pictures to push notifications (William Casarin)
-
-
-### Changed
-
-- Don't show reposts for the same note more than once in your home feed (William Casarin)
-- Improved profile image bandwidth optimization (Daniel D’Aquino)
-- Improved reliability of picture selector (Daniel D’Aquino)
-- Changed spaces to newlines in new posts to provide cleaner separation between text, uploaded media, and quoted notes (Terry Yiu)
-
-
-### Fixed
-
-- Fixed issue where some push notifications would not open in the app and leave users confused (Daniel D’Aquino)
-- Fixed issue where app would need a restart for new NWC wallets to work (Daniel D’Aquino)
-- Fixed overly sensitive horizontal swipe on thread chat view (Daniel D’Aquino)
-- Trim whitespaces from Lightning addresses (Terry Yiu)
-- Fixed translation export script by upgrading nostr-sdk-swift dependency to support Mac Catalyst (Terry Yiu)
-- Fixed issue where users continue to receive push notifications after logout (Daniel D’Aquino)
-- Fixed an issue where events on a thread view would occasionally disappear (Daniel D’Aquino)
-- Improved robustness of the URL handler (Daniel D’Aquino)
-- Translate notes even if they are in a preferred language but not the current language as that is what users expect (Terry Yiu)
-- Cancel ongoing uploading operations after the user cancels the post (Swift Coder)
-- Fixed link and photo sharing support on macOS (Swift Coder)
-- Fix bug where profile view was showing more than just the notes and replies on the notes / notes & replies tabs (Terry Yiu)
-- Fixed reposts banner to be localizable (Terry Yiu)
-
-
-### Removed
-
-- Removed language filtering from Universe feed because language detection can be inaccurate (Terry Yiu)
-- Removed mystery tabs meant to fix tab switching bug that no longer exists (Terry Yiu)
-
-
-
-[1.13](https://github.com/damus-io/damus/releases/tag/v1.13): https://github.com/damus-io/damus/releases/tag/v1.13
-
-
 ## [1.12.3] - 2025-02-06
 
 ### Added
