+++ damus/hu-HU.lproj/Localizable.stringsdict
@@ -66,22 +66,6 @@
 			<string>Importok</string>
 		</dict>
 	</dict>
-	<key>people_reposted_count</key>
-	<dict>
-		<key>NSStringLocalizedFormatKey</key>
-		<string>%#@REPOSTED@</string>
-		<key>REPOSTED</key>
-		<dict>
-			<key>NSStringFormatSpecTypeKey</key>
-			<string>NSStringPluralRuleType</string>
-			<key>NSStringFormatValueTypeKey</key>
-			<string>d</string>
-			<key>one</key>
-			<string>%2$@ és %1$d egyéb újraosztva</string>
-			<key>other</key>
-			<string>%2$@ és %1$d egyebek újraosztva</string>
-		</dict>
-	</dict>
 	<key>reacted_tagged_in_3</key>
 	<dict>
 		<key>NSStringLocalizedFormatKey</key>
